+++ docs/tr2/CHANGELOG.md
@@ -18,6 +18,7 @@
 - added the ability to change the health bar color for allies, defaulting to green (#3005)
 - added the ability to skip consecutive credit images by holding the action / escape keys
 - added the ability to cycle between a list of predefined Lara poses in the photo mode
+- added the ability to use the dev console during FMVs
 - added a new easter egg command
 - added a `/lighting` console command to let the player turn lighting system on/off
 - added an `/immune` console command to make Lara impervious to damage